[#22] refactor : 쿠폰 controller 역할에 맞게 클래스 분리 리팩토링
