+++ src/main/java/com/flab/goodchoice/coupon/api/CouponIssueInfoController.java
@@ -1,0 +1,26 @@
+package com.flab.goodchoice.coupon.api;
+
+import com.flab.goodchoice.coupon.application.CouponIssuanceService;
+import com.flab.goodchoice.coupon.dto.MemberSpecificCouponResponse;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RequestMapping("/api/coupons")
+@RestController
+public class CouponIssueInfoController {
+
+    private final CouponIssuanceService couponIssuanceService;
+
+    public CouponIssueInfoController(CouponIssuanceService couponIssuanceService) {
+        this.couponIssuanceService = couponIssuanceService;
+    }
+
+    @GetMapping("/publish/{memberId}")
+    public List<MemberSpecificCouponResponse> getMemberCoupon(@PathVariable Long memberId) {
+        return couponIssuanceService.getMemberCoupon(memberId);
+    }
+}
