+++ src/main/java/com/flab/goodchoice/coupon/application/CouponIssuanceService.java
@@ -3,11 +3,12 @@ package com.flab.goodchoice.coupon.application;
 import com.flab.goodchoice.common.aop.LimitedCountLock;
 import com.flab.goodchoice.coupon.domain.Coupon;
 import com.flab.goodchoice.coupon.domain.CouponPublish;
-import com.flab.goodchoice.coupon.domain.Member;
 import com.flab.goodchoice.coupon.dto.MemberSpecificCouponResponse;
 import com.flab.goodchoice.coupon.exception.CouponError;
 import com.flab.goodchoice.coupon.exception.CouponException;
 import com.flab.goodchoice.coupon.infrastructure.repositories.AppliedUserRepository;
+import com.flab.goodchoice.member.application.MemberQuery;
+import com.flab.goodchoice.member.domain.model.Member;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
