+++ src/test/java/com/flab/goodchoice/coupon/application/CouponPublishServiceTest.java
@@ -2,14 +2,17 @@ package com.flab.goodchoice.coupon.application;
 
 import com.flab.goodchoice.coupon.domain.Coupon;
 import com.flab.goodchoice.coupon.domain.CouponType;
-import com.flab.goodchoice.coupon.domain.Member;
 import com.flab.goodchoice.coupon.domain.State;
 import com.flab.goodchoice.coupon.dto.MemberSpecificCouponResponse;
 import com.flab.goodchoice.coupon.exception.CouponException;
-import com.flab.goodchoice.coupon.exception.MemberException;
 import com.flab.goodchoice.coupon.infrastructure.*;
 import com.flab.goodchoice.coupon.infrastructure.entity.CouponEntity;
 import com.flab.goodchoice.coupon.infrastructure.repositories.*;
+import com.flab.goodchoice.member.application.MemberCommand;
+import com.flab.goodchoice.member.application.MemberQuery;
+import com.flab.goodchoice.member.domain.model.Member;
+import com.flab.goodchoice.member.domain.repositories.MemberRepository;
+import com.flab.goodchoice.member.exception.MemberException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
